Dropped unsafe code
A custom linked list in the algorithm that computes connected components
has been replaced with a hash map. In my back-of-the-envelope benchmarks
the hash map is slower and uses more memory than a linked list, but
these differences are negligible in comparison to the overal run time
and memory usage.